Revert "refactor: rename file to accurately reflect RFC 327 PR number"
This reverts commit bb91709e030c73875b502670349f23539976f80d.
